Add Cultureinfo.invariantculture to input amount. change createAccount to take decimal as parameter instead of int

+++ LifeCycleBank/Bank.cs
@@ -100,7 +100,7 @@ namespace LifeCycleBank
             }
         }
 
-        public string CreateAccount(ICustomer customerId, int balance)
+        public string CreateAccount(ICustomer customerId, decimal balance)
         {
             try
             {
